// src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Parse string date to type Date */
  CacheDate: any;
  /** The javascript `Date` as string. Type represents date and time as the ISO Date string. */
  DateTime: any;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: any;
};
// ...
export type Mutation = {
// ...
  changeSettingsProfileNotifications: SettingsProfileNotifications;
// ...
};
// ...
export type MutationChangeSettingsProfileNotificationsArgs = {
  data: UpdateSettingsProfileNotificationInput;
};
// ...
export type Query = {
// ...
  getSettingsProfileNotifications: SettingsProfileNotifications;
// ...
};
// ...
export type QueryGetSettingsProfileNotificationsArgs = {
  profile_id: Scalars['String'];
};
// ...
export type SettingsProfileNotifications = {
  __typename?: 'SettingsProfileNotifications';
  _id: Scalars['ID'];
  profile_id: Scalars['String'];
  comments_push: Scalars['Boolean'];
  comments_email: Scalars['Boolean'];
  comment_response_push: Scalars['Boolean'];
  comment_response_email: Scalars['Boolean'];
  comment_claps_push: Scalars['Boolean'];
  comment_claps_email: Scalars['Boolean'];
  claps_push: Scalars['Boolean'];
  claps_email: Scalars['Boolean'];
  messages_received_push: Scalars['Boolean'];
  messages_received_email: Scalars['Boolean'];
  message_request_push: Scalars['Boolean'];
  message_request_email: Scalars['Boolean'];
  followers_request_push: Scalars['Boolean'];
  followers_request_email: Scalars['Boolean'];
  follower_new_push: Scalars['Boolean'];
  follower_new_email: Scalars['Boolean'];
  mentions_push: Scalars['Boolean'];
  mentions_email: Scalars['Boolean'];
  challenge_reminder_push: Scalars['Boolean'];
  challenge_reminder_email: Scalars['Boolean'];
  challenge_subscribed_received_push: Scalars['Boolean'];
  challenge_subscribed_received_email: Scalars['Boolean'];
  maintenance_reminder_push: Scalars['Boolean'];
  maintenance_reminder_email: Scalars['Boolean'];
  activities_received_push: Scalars['Boolean'];
  activities_received_email: Scalars['Boolean'];
};
// ...
export type UpdateSettingsProfileNotificationInput = {
  profile_id: Scalars['String'];
  comments_push?: Maybe<Scalars['Boolean']>;
  comments_email?: Maybe<Scalars['Boolean']>;
  comment_response_push?: Maybe<Scalars['Boolean']>;
  comment_response_email?: Maybe<Scalars['Boolean']>;
  comment_claps_push?: Maybe<Scalars['Boolean']>;
  comment_claps_email?: Maybe<Scalars['Boolean']>;
  claps_push?: Maybe<Scalars['Boolean']>;
  claps_email?: Maybe<Scalars['Boolean']>;
  messages_received_push?: Maybe<Scalars['Boolean']>;
  messages_received_email?: Maybe<Scalars['Boolean']>;
  message_request_push?: Maybe<Scalars['Boolean']>;
  message_request_email?: Maybe<Scalars['Boolean']>;
  followers_request_push?: Maybe<Scalars['Boolean']>;
  followers_request_email?: Maybe<Scalars['Boolean']>;
  follower_new_push?: Maybe<Scalars['Boolean']>;
  follower_new_email?: Maybe<Scalars['Boolean']>;
  mentions_push?: Maybe<Scalars['Boolean']>;
  mentions_email?: Maybe<Scalars['Boolean']>;
  challenge_reminder_push?: Maybe<Scalars['Boolean']>;
  challenge_reminder_email?: Maybe<Scalars['Boolean']>;
  challenge_subscribed_received_push?: Maybe<Scalars['Boolean']>;
  challenge_subscribed_received_email?: Maybe<Scalars['Boolean']>;
  maintenance_reminder_push?: Maybe<Scalars['Boolean']>;
  maintenance_reminder_email?: Maybe<Scalars['Boolean']>;
  activities_received_push?: Maybe<Scalars['Boolean']>;
  activities_received_email?: Maybe<Scalars['Boolean']>;
};